client/requirements.txt
mysql-connector-python==8.0.30
python-dotenv==0.20.0
loguru==0.6.0